dom/tests/mochitest/script/test_bug1053321.html

<!DOCTYPE html>
<html>
<!--
This test confirms we don't block body referred sub-resources by head-referenced defer and async scripts.
If this test times out, the two image requests, that unblock the two script requests, never happen, hence
are unexpectedly blocked.
-->

<head>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <!-- this script is not loaded until file_blocked_script.sjs?unblock&defer request is made,
       when this script is executed, it sets window.script_executed_defer to true
   -->
  <script defer src="file_blocked_script.sjs?blocked&defer"></script>

  <!-- this script is not loaded until file_blocked_script.sjs?unblock&async request is made,
       when this script is executed, it sets window.script_executed_async to true
   -->
  <script async src="file_blocked_script.sjs?blocked&async"></script>
</head>

<body>
  <script>
    SimpleTest.waitForExplicitFinish();

    // We can't test whether the two scripts have not been executed here, since
    // preloads of the two images below (that unblock the two tested <head>
    // scripts) may happen sooner than this script executes.
    document.addEventListener("DOMContentLoaded", function() {
      ok(window.script_executed_defer, "Deferred script executed before DOMContentLoaded");
    });
    window.addEventListener("load", function() {
      ok(window.script_executed_async, "Async script executed before onload");
      SimpleTest.finish();
    }, true);
  </script>
  <img src="file_blocked_script.sjs?unblock&defer"/>
  <img src="file_blocked_script.sjs?unblock&async"/>
</body>
